+++ pdfminer/data_structures.py
@@ -45,7 +45,7 @@ class NumberTree:
         values = self._parse()
 
         if settings.STRICT:
-            if not all(a[0] <= b[0] for a, b in zip(values, values[1:])):
+            if not all(a[0] <= b[0] for a, b in zip(values, values[1:], strict=False)):
                 raise PDFSyntaxError("Number tree elements are out of order")
         else:
             values.sort(key=lambda t: t[0])