+++ pdfminer/image.py
@@ -153,8 +153,8 @@ class ImageWriter:
             if LITERAL_DEVICE_CMYK in image.colorspace:
                 try:
                     from PIL import Image, ImageChops  # type: ignore[import]
-                except ImportError:
-                    raise ImportError(PIL_ERROR_MESSAGE)
+                except ImportError as err:
+                    raise ImportError(PIL_ERROR_MESSAGE) from err
 
                 ifp = BytesIO(data)
                 i = Image.open(ifp)
@@ -174,8 +174,8 @@ class ImageWriter:
         with open(path, "wb") as fp:
             try:
                 from PIL import Image  # type: ignore[import]
-            except ImportError:
-                raise ImportError(PIL_ERROR_MESSAGE)
+            except ImportError as err:
+                raise ImportError(PIL_ERROR_MESSAGE) from err
 
             # if we just write the raw data, most image programs
             # that I have tried cannot open the file. However,
@@ -245,8 +245,8 @@ class ImageWriter:
                     Image,  # type: ignore[import]
                     ImageOps,
                 )
-            except ImportError:
-                raise ImportError(PIL_ERROR_MESSAGE)
+            except ImportError as err:
+                raise ImportError(PIL_ERROR_MESSAGE) from err
 
             mode: Literal["1", "L", "RGB", "CMYK"]
             if image.bits == 1:
@@ -278,7 +278,7 @@ class ImageWriter:
     @staticmethod
     def _is_jbig2_iamge(image: LTImage) -> bool:
         filters = image.stream.get_filters()
-        for filter_name, params in filters:
+        for filter_name, _params in filters:
             if filter_name in LITERALS_JBIG2_DECODE:
                 return True
         return False