+++ pdfminer/jbig2.py
@@ -138,7 +138,7 @@ class JBIG2StreamReader:
             ref_count = unpack_int(">L", field)
             ref_count = masked_value(REF_COUNT_LONG_MASK, ref_count)
             ret_bytes_count = int(math.ceil((ref_count + 1) / 8))
-            for ret_byte_index in range(ret_bytes_count):
+            for _ret_byte_index in range(ret_bytes_count):
                 ret_byte = unpack_int(">B", self.stream.read(1))
                 for bit_pos in range(7):
                     retain_segments.append(bit_set(bit_pos, ret_byte))
@@ -154,7 +154,7 @@ class JBIG2StreamReader:
 
         ref_size = calcsize(ref_format)
 
-        for ref_index in range(ref_count):
+        for _ref_index in range(ref_count):
             ref_data = self.stream.read(ref_size)
             ref = unpack_int(ref_format, ref_data)
             ref_segments.append(ref)