+++ tools/dumppdf.py
@@ -180,7 +180,7 @@ def dumpoutline(
     try:
         outlines = doc.get_outlines()
         outfp.write("<outlines>\n")
-        for level, title, dest, a, se in outlines:
+        for level, title, dest, a, _se in outlines:
             pageno = None
             if dest:
                 dest = resolve_dest(dest)